refactor(apollo): rename CheckboxText component folder
